+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/mapping/FlagMapping.kt
@@ -29,7 +29,7 @@ internal fun DownstreamFlagDto.toDomain(): Flag {
         createdAt = approved_at,
         updatedAt = updated_at,
         reviewedAt = reviewed_at,
-        approvedAt = reviewed_by,
+        approvedAt = approved_at,
         rejectedAt = rejected_at,
     )
 }