+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/channels/UnreadCountIndicator.kt
@@ -44,7 +44,8 @@ public fun UnreadCountIndicator(
     modifier: Modifier = Modifier,
     color: Color = ChatTheme.colors.errorAccent,
 ) {
-    val displayText = if (unreadCount > 99) UNREAD_COUNT_MANY else unreadCount.toString()
+
+    val displayText = if (unreadCount > LimitTooManyUnreadCount) UnreadCountMany else unreadCount.toString()
     val shape = RoundedCornerShape(9.dp)
 
     Box(
@@ -66,7 +67,8 @@ public fun UnreadCountIndicator(
 /**
  * The unread count that's shown for many messages.
  */
-private const val UNREAD_COUNT_MANY = "99+"
+private const val UnreadCountMany = "99+"
+private const val LimitTooManyUnreadCount = 99
 
 /**
  * Preview of [UnreadCountIndicator] with few unread messages.